ebay_mock/requirements.txt
amqp==5.1.1
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
boltons==21.0.0
bracex==2.3.post1
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
dataclasses==0.6
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.2
django-cors-headers==4.1.0
django-enumchoicefield==3.0.1
django-extensions==3.2.3
django-redis==5.3.0
django-shell==1.0.2
djangorestframework==3.14.0
drf-spectacular==0.26.3
ebaysdk==2.2.0
executing==1.2.0
face==22.0.0
fastjsonschema==2.17.1
filelock==3.12.2
fqdn==1.5.1
glom==22.1.0
html2text==2020.1.16
identify==2.5.24
idna==3.4
inflection==0.5.1
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-contrib-core==0.4.2
jupyter-contrib-nbextensions==0.7.0
jupyter-events==0.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-nbextensions-configurator==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kombu==5.3.1
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nodeenv==1.8.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.0
overrides==7.3.1
packaging==23.1
pandas==2.0.2
pandocfilters==1.5.0
parso==0.8.3
peewee==3.16.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.6.0
pre-commit==3.3.3
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-jsonrpc==1.0.0
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
redis==4.5.5
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.4.2
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
semgrep==1.27.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.3
tzdata==2023.3
ujson==5.8.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
virtualenv==20.23.1
wcmatch==8.4.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.7
